ticketlist is now ticketList.  All url based paramaters will be camel cased. ticketlist template is now called generic. This template was so basic that it really did not need to be different.
renamed display_ticket.php to get_ticket.php to be more inline with the naming convention I tend to use.

updated the db connection object to the latest version

The start of user pages is in.  right now it only shows information from their user table.  It will be expanded with a lot more cool things.