Live Ticket Update XXI v.3.1.0.r314 Bug Fixes Fixed bug #545. Fixed bug #693.
New Features
Creating a new ticket, or a new bug report now shows up inside of the ticket display instead of as a popup.
Adding a reply now follows the same look and feel of new tickets and bug reports.

Known Issues
Searching By project does not work
There are no page numbers, nor any way to show more then the default amount of tickets in the list view.
There are some issues with editing that have been around since editing was in place. All of the ticket code links are not working as intended with the new editing system.
Closed tickets are not showing up for some people. Even though a number shows up next to Closed Tickets.