Ticket Update XXIV v.3.6.0.r730 Bug Fixes Fixed a 404 error when trying to link to a js library Fixed OpenDNS Server 1/PING is WARNING
New Features
links in the form of http://www.google.com will transform into clickable links
There is a new admin section for users who have access to it. It is listed on the tab bar. If you see it you have accessWill be included in the next update
Ticket list is now sortable Will be included in the next update
Users are now given a http://www.gravatar.com Icon

Known Issues
Searching By project does not work
Page indicators in a search result do not currently work.
There are some issues with editing that have been around since editing was in place. All of the ticket code links are not working as intended with the new editing system.
Commenting on a ticket that was closed or reopened will generate extra replies